update login and signup forms to use react-hook-form
